Add files via upload
Plan_Utils is a toolbox containing all of the functions needed to automated IGRT matching. Of particular interest is the find_GTV_overlap_vector which can take a base rtss and adaptive rtss path (for two rtss dicom files corresponding to base plan and adaptive place), an initial guess (can use [0,0,0]) of the shift vector, and the two names of the contours (in each plan) to mazimize overlap. This returns a shift vector that when applied to the adaptive structure set, maximizes overlap between the two named contours in oarbase and oardata.