Add files via upload
This is an example deployment of Plan_Utils. This script assumes it is run in a directory containing N sub directories, where N is the number of patients. All N sub directories have names starting with 'AU_PA' and this is assumed by the script (modifiable). It also assumes that each patient sub directory itself contains a folder called 'TP0' corresponding to the base plan, and containing the base plan RT-STRUCT.dcm file (called 'RT-STRUCT.dcm', name modifiable). It also assumed that each patient sub directory contains the 'RT-STRUCT' for all fractions in folders called TP1, TP2, TP3...etc. corresponding to the fractions.

The script then finds the shift vector for IGRT matching for all of these, and then extracts DVH features after IGRT matching is done. It saves the features and shift vectors for all patients in their sub-directory in the form of a csv file.